+++ epp/__init__.py
@@ -1,1 +1,6 @@
-from .core import EPP, Contact, Domain, Nameserver
+# -*- coding: utf-8 -*-
+
+from .contact import Contact
+from .domain import Domain
+from .epp import EPP
+from .nameserver import Nameserver
